refine

+++ designer/server/src/test/scala/pl/touk/nussknacker/ui/process/test/PreliminaryScenarioTestDataSerDeTest.scala
@@ -18,7 +18,7 @@ class PreliminaryScenarioTestDataSerDeTest extends AnyFunSuite with Matchers wit
     TestDataSettings(
       maxSamplesCount = maxSamplesCount,
       testDataMaxLength = testDataMaxLength,
-      resultsMaxBytes = Int.MaxValue
+      resultsMaxBytes = Long.MaxValue
     )
   )
